+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to
 
 ## [Unreleased]
 
+- Added a `typed_header` method to the `ClientRequest` and `RequestBuilderExt`
+  for typed header insertion.
+
 ## [0.5.0] - 2025.03.31
 
 - **breaking** `body_reader` implementation has been replaced with the